# f1api/openapi/api.yml
openapi: "3.0.0"
info:
  title: F1 API
  version: "1.0.0"
  description: |
    ## Paging and Sorting

    The following query paramters can be used to control the paging and sorting of collections.

    ### `pageSize`

    Defines the number of elements per page.

    **Default value:** 25

    **Notice:** maximum number is 100

    ### `pageNumber`

    Specifies the page number that should be returned.

    **Notice:** first page has the number 0

    **Default value:** 0

    ### `sort`

    Specifies by which properties to sort. <br>
    By default the sort order is ascending. This can be changed by adding `asc` for ascending or `desc` for
    descending separated by a comma after the sort property.

    **Hint:** To sort by more than one property, simply add a new sort parameter to the query.

    ### Example

    GET `/drivers?pageSize=20&pageNumber=0&sort=firstName,asc`

    will return the first page with a maximum of 20 drivers sorted by first name in ascending order.
...
paths:
...
  /engines:
    post:
      x-internal: true
      summary: Create engine
      tags:
        - "Engines"
        - "EngineController"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Engine"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engine"
        400:
          $ref: "./components/responses.yml#/400"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get engines / Get engine by manufacturer
      tags:
        - "Engines"
        - "EngineController"
      parameters:
        - $ref: "./components/parameters.yml#/manufacturer"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engines"
        400:
          $ref: "./components/responses.yml#/400"
        500:
          $ref: "./components/responses.yml#/500"
  /engines/{engineId}:
    get:
      summary: Get engine
      tags:
        - "Engines"
        - "EngineController"
      parameters:
        - $ref: "./components/parameters.yml#/engineId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engine"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update engine
      tags:
        - "Engines"
        - "EngineController"
      parameters:
        - $ref: "./components/parameters.yml#/engineId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Engine"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Engine"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /engines/{engineId}/teams:
    get:
      summary: Get all teams which used the engine
      tags:
        - "Teams"
        - "EngineController"
      parameters:
        - $ref: "./components/parameters.yml#/engineId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Teams"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
  /engines/{engineId}/seasons:
    get:
      summary: Get all seasons in that the engine was used
      tags:
        - "Seasons"
        - "EngineController"
      parameters:
        - $ref: "./components/parameters.yml#/engineId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Seasons"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"

  # TeamOfSeasonController
...